pyproject.toml
[tool.poetry]
name = "world-boss-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "world_boss"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = {extras = ["async"], version = "^2.2.2"}
mypy = "^0.991"
flask-sqlalchemy = "2.5.1"
flask-migrate = "^4.0.0"
psycopg2 = "^2.9.5"
redis = "^4.3.5"
gunicorn = "^20.1.0"
types-flask-migrate = "^4.0.0.0"
types-redis = "^4.3.21.6"
boto3 = "^1.26.22"
ethereum-kms-signer = "^0.1.6"
pyasn1 = "^0.4.8"
backoff = "^2.2.1"
types-urllib3 = "^1.26.25.4"
httpx = "^0.23.1"
slack-sdk = "^3.19.5"
celery = "^5.2.7"
celery-types = "^0.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-flask-sqlalchemy = "^1.1.0"
pytest-postgresql = "^4.1.1"
pytest-dotenv = "^0.5.2"
psycopg-binary = "^3.1.4"
psycopg = "^3.1.4"
pytest-redis = "^2.4.0"
pre-commit = "^2.20.0"
black = "^22.10.0"
bencodex = "^1.0.1"
pytest-httpx = "^0.21.2"
pytest-asyncio = "^0.20.2"
pytest-timer = "^0.0.11"
pytest-celery = "^0.0.0"

[tool.pytest.ini_options]
env_files = [".test.env", ".local.env"]
plugins = ["pytest-flask-sqlalchemy"]
addopts = "--basetemp=/tmp/pytest --redis-exec=/opt/homebrew/bin/redis-server"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "migrations"
